redux: Move connect preserve-on-back helpers out of misc.js.
`connect` is a Redux thing, so it makes sense to store helpers
for `connect` in redux.js. It's also good to have as few functions
as possible in a generically-named file like misc.js.
